+++ src/getDeploy/utils.ts
@@ -1,14 +1,21 @@
 import {
   getModuleData,
   type GetModuleData,
+  type GetModuleNameByType,
   type ModuleName,
 } from '@inverter-network/abis'
-import { getContract } from 'viem'
+import { getContract, parseUnits } from 'viem'
 import { METADATA_URL, DEPLOYMENTS_URL } from './constants'
 import { ERC20_ABI } from '../utils/constants'
 
 import type { PublicClient, WalletClient } from 'viem'
-import type { FactoryType, UserModuleArg } from '..'
+import type {
+  FactoryType,
+  FilterByPrefix,
+  GetUserArgs,
+  PopWalletClient,
+  UserModuleArg,
+} from '..'
 import type { Abi } from 'abitype'
 
 type DeploymentResponse = {
@@ -16,6 +23,7 @@ type DeploymentResponse = {
   erc20Mock: Record<string, `0x${string}` | undefined>
   orchestratorFactory: Record<string, `0x${string}` | undefined>
   restrictedPimFactory: Record<string, `0x${string}` | undefined>
+  immutablePimFactory: Record<string, `0x${string}` | undefined>
 }
 
 export type GetViemMethodsParams<FT extends FactoryType> = {
@@ -32,6 +40,49 @@ export const fetchDeployment = async (
 ): Promise<DeploymentResponse> => {
   const response = await fetch(`${DEPLOYMENTS_URL}/${version}.json`)
   return await response.json()
+}
+
+export const getAbi = <FT extends FactoryType>(factoryType: FT) => {
+  const abi = getModuleData(
+    (() => {
+      switch (factoryType) {
+        case 'default':
+          return 'OrchestratorFactory_v1'
+        case 'restricted-pim':
+          return 'Restricted_PIM_Factory_v1'
+        case 'immutable-pim':
+          return 'Immutable_PIM_Factory_v1'
+        default:
+          throw new Error('Unsupported factory type')
+      }
+    })()
+  ).abi as GetModuleData<
+    FT extends 'restricted-pim'
+      ? 'Restricted_PIM_Factory_v1'
+      : FT extends 'immutable-pim'
+        ? 'Immutable_PIM_Factory_v1'
+        : 'OrchestratorFactory_v1'
+  >['abi']
+
+  return abi
+}
+
+export const getMethodName = <FT extends FactoryType>(factoryType: FT) => {
+  const methodName = (() => {
+    switch (factoryType) {
+      case 'default':
+        return 'createOrchestrator'
+      case 'restricted-pim':
+      case 'immutable-pim':
+        return 'createPIMWorkflow'
+      default:
+        throw new Error('Unsupported factory type')
+    }
+  })() as FT extends 'restricted-pim' | 'immutable-pim'
+    ? 'createPIMWorkflow'
+    : 'createOrchestrator'
+
+  return methodName
 }
 
 // retrieves the deployment function via viem
@@ -61,6 +112,8 @@ export const getViemMethods = async ({
             return 'orchestratorFactory'
           case 'restricted-pim':
             return 'restrictedPimFactory'
+          case 'immutable-pim':
+            return 'immutablePimFactory'
           default:
             throw new Error('Unsupported factory type')
         }
@@ -70,27 +123,19 @@ export const getViemMethods = async ({
   if (!address)
     throw new Error('Chain ID is not supported @ deployment factory address')
 
+  const methodName = getMethodName(factoryType)
+
   const { write, simulate, estimateGas } = getContract({
     address,
-    abi,
+    abi: abi as Abi,
     client: {
       wallet: walletClient,
       public: publicClient,
     },
   })
 
-  const methodName = (() => {
-    switch (factoryType) {
-      case 'default':
-        return 'createOrchestrator'
-      case 'restricted-pim':
-        return 'createPIMWorkflow'
-      default:
-        throw new Error('Unsupported factory type')
-    }
-  })()
-
   return {
+    factoryAddress: address,
     simulateWrite: simulate[methodName],
     write: write[methodName],
     estimateGas: estimateGas[methodName],
@@ -139,23 +184,87 @@ export const getDefaultToken = async (
   return { defaultToken: tokenAddress!, decimals }
 }
 
-export const getAbi = <FT extends FactoryType>(factoryType: FT) => {
-  const abi = getModuleData(
-    (() => {
-      switch (factoryType) {
-        case 'default':
-          return 'OrchestratorFactory_v1'
-        case 'restricted-pim':
-          return 'Restricted_PIM_Factory_v1'
-        default:
-          throw new Error('Unsupported factory type')
-      }
-    })()
-  ).abi as GetModuleData<
-    FT extends 'restricted-pim'
-      ? 'Restricted_PIM_Factory_v1'
-      : 'OrchestratorFactory_v1'
-  >['abi']
+const isPimArgs = (
+  args: any,
+  factoryType: FactoryType
+): args is GetUserArgs<
+  {
+    fundingManager: FilterByPrefix<
+      GetModuleNameByType<'fundingManager'>,
+      'FM_BC'
+    >
+    paymentProcessor: GetModuleNameByType<'paymentProcessor'>
+    authorizer: GetModuleNameByType<'authorizer'>
+  },
+  'restricted-pim' | 'immutable-pim'
+> => {
+  switch (factoryType) {
+    case 'restricted-pim':
+      return !!args?.fundingManager?.bondingCurveParams?.initialCollateralSupply
+    case 'immutable-pim':
+      return !!args?.initialPurchaseAmount
+    default:
+      return false
+  }
+}
 
-  return abi
+export const handlePimFactoryApprove = async ({
+  factoryType,
+  factoryAddress,
+  userArgs,
+  walletClient,
+  publicClient,
+}: {
+  userArgs: any
+  factoryType: FactoryType
+  factoryAddress: `0x${string}`
+  walletClient: PopWalletClient
+  publicClient: PublicClient
+}) => {
+  const handle = async (requiredAllowance: string) => {
+    // get the collateral token address
+    const collateralTokenAddress = userArgs.fundingManager.collateralToken
+    // get the ERC20 contract
+    const contract = getContract({
+      address: collateralTokenAddress,
+      abi: ERC20_ABI,
+      client: { wallet: walletClient, public: publicClient },
+    })
+    // get the decimals of the ERC20 token
+    const decimals = await contract.read.decimals()
+    // parse the required allowance
+    const parsedRequiredAllowance = parseUnits(requiredAllowance, decimals)
+    // get the current allowance
+    const allowance = await contract.read.allowance([
+      walletClient.account.address,
+      factoryAddress,
+    ])
+    // check if the current allowance is enough
+    const hasEnoughAllowance = allowance >= parsedRequiredAllowance
+    // if the current allowance is enough, return
+    if (hasEnoughAllowance) return
+    // if the current allowance is not enough, approve the factory
+    const hash = await contract.write.approve([
+      factoryAddress,
+      parsedRequiredAllowance,
+    ])
+    // wait for the transaction to be mined
+    await publicClient.waitForTransactionReceipt({ hash })
+  }
+
+  // check if the userArgs are for a PIM factory
+  if (!isPimArgs(userArgs, factoryType)) return
+  // handle the PIM factory approve
+  switch (factoryType) {
+    case 'restricted-pim':
+      await handle(
+        userArgs.fundingManager.bondingCurveParams.initialCollateralSupply
+      )
+      break
+    case 'immutable-pim':
+      await handle(userArgs.initialPurchaseAmount)
+      break
+  }
+
+  return
 }