+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.2.2](https://github.com/InverterNetwork/sdk/compare/v0.2.1...v0.2.2) (2024-08-23)
+
+
+### Features
+
+* immutable pim factory ([c811a17](https://github.com/InverterNetwork/sdk/commit/c811a17e86a8a2327215708ed7abf376a9834bd6))
+
+### [0.2.1](https://github.com/InverterNetwork/sdk/compare/v0.2.0...v0.2.1) (2024-08-22)
+
+
+### Features
+
+* auto allowance at pim-factory ([2303f2d](https://github.com/InverterNetwork/sdk/commit/2303f2d539cc3e1892025b1369869a3914477e5f))
+
+## [0.2.0](https://github.com/InverterNetwork/sdk/compare/v0.2.0-alpha.6...v0.2.0) (2024-08-21)
+
 ### [0.1.8](https://github.com/InverterNetwork/sdk/compare/v0.1.8-alpha.0...v0.1.8) (2024-08-12)
 
 ### [0.1.7](https://github.com/InverterNetwork/sdk/compare/v0.1.7-alpha.2...v0.1.7) (2024-08-07)